Heroes and problems classes implemented

+++ src/main/java/agh/ics/oop/IProblem.java
@@ -3,5 +3,5 @@ package agh.ics.oop;
 public interface IProblem {
     int getSolvingTime();
     int getDestructionTime();
-    float getOccuranceChance();
+    double getOccuranceChance();
 }
